[cheri] Increase the inlined frame limit (24->48)
Double the inlined frame limit to account for capabilities being double
the width of the pointer size on the native architecture. This ensures
the balance between the cost of inlining and referencing the frame with
a pointer remains broadly similar to that on non-CHERI architectures
(without this change all frames are referenced with a pointer and non
are inlined).